Moved error into error pkg
