+++ conf/script/src/build_system/cmd/hierarchy/find_build_dir/cli.py
@@ -3,7 +3,7 @@ from typing import Final
 
 import utils.cli.arg
 import utils.cli.arg_parsing
-import utils.cli.error
+import utils.error.error
 import utils.error.status
 from build_system import cmd
 from build_system.cmd.hierarchy.find_build_dir.find_build_dir import BUILD_DIR_NAME
@@ -22,10 +22,10 @@ def find_build_dir():
         build_dir = cmd.hierarchy.find_build_dir.find_build_dir()
         print(build_dir, end=str())
 
-    except utils.cli.error.RootDirNotFoundError as raised_exception:
+    except utils.error.error.RootDirNotFoundError as raised_exception:
         arg_parser.exit(utils.error.status.ErrorStatus.ROOT_DIR_NOT_FOUND, str(raised_exception))
 
-    except utils.cli.error.BuildDirNotFoundError as raised_exception:
+    except utils.error.error.BuildDirNotFoundError as raised_exception:
         arg_parser.exit(utils.error.status.ErrorStatus.BUILD_DIR_NOT_FOUND, str(raised_exception))
 
     except OSError as raised_exception:
