+++ conf/script/src/utils/error/synthesis.py
@@ -1,17 +1,17 @@
 from typing import Final, Type
 
-import utils.cli.error
+import utils.error.error
 import utils.error.status
 import utils.error.manage
 
 ALL_ERRORS: Final[list[Type[utils.cli.managed_error.ManagedError]]] = [
-    utils.cli.error.SuccessWarning,
-    utils.cli.error.UnsupportedError,
-    utils.cli.error.ArgParserError,
-    utils.cli.error.UnknownParsedArgError,
-    utils.cli.error.EmptyParsedArgError,
-    utils.cli.error.RootDirNotFoundError,
-    utils.cli.error.BuildDirNotFoundError,
+    utils.error.error.SuccessWarning,
+    utils.error.error.UnsupportedError,
+    utils.error.error.ArgParserError,
+    utils.error.error.UnknownParsedArgError,
+    utils.error.error.EmptyParsedArgError,
+    utils.error.error.RootDirNotFoundError,
+    utils.error.error.BuildDirNotFoundError,
 ]
 
 ALL_ERRORS_BY_STATUS: Final[dict[utils.error.status.ErrorStatus, Type[utils.cli.managed_error.ManagedError]]] = {status: ALL_ERRORS[status] for status in
